[frontend] refactor WorkflowStatus
