+++ opencti-platform/opencti-front/src/private/components/common/workflow/WorkflowStatus.test.tsx
@@ -3,7 +3,7 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { screen, waitFor } from '@testing-library/react';
 import WorkflowStatus, { WorkflowTransitions } from './WorkflowStatus';
 import testRender from '../../../../utils/tests/test-render';
-import type { WorkflowStatus_data$key } from './__generated__/WorkflowStatus_data.graphql';
+import type { workflowStatus_data$key } from './__generated__/workflowStatus_data.graphql';
 import { CommentMode } from '../../settings/sub_types/workflow/utils';
 
 // ---------------------------------------------------------------------------
@@ -54,7 +54,7 @@ const makeStatus = (color = '#ff0000', name = 'In review') => ({
   template: { name, color },
 });
 
-const makeDraft = (overrides: Record<string, unknown> = {}): WorkflowStatus_data$key => ({
+const makeDraft = (overrides: Record<string, unknown> = {}): workflowStatus_data$key => ({
   id: 'draft-1',
   entity_id: 'entity-1',
   processingCount: 0,
@@ -66,7 +66,7 @@ const makeDraft = (overrides: Record<string, unknown> = {}): WorkflowStatus_data
     allowedTransitions: [],
   },
   ...overrides,
-} as unknown as WorkflowStatus_data$key);
+} as unknown as workflowStatus_data$key);
 
 const makeTransition = (overrides: Record<string, unknown> = {}) => ({
   event: 'approve',
